Add the gRPC client and queue processing
Add the gRPC client with only the required proto files. This decision
theoretically reduces the amount of updates we would need to do if the
gRPC definitions update but this may come back to bite us if it turns
out it is just easier copying files across the repos. We will have to
see.

In principle we should really move the `RpcClient` to the new project
and rename it to `GrpcClient` but that then means faffing about with
shared projects for the interfaces etc. which is Overkill seeing as this
will probably be the only ever implementation. If it turns out that we
do implement another client (For example, a Tacview one) then we will
do the shared project incantations at that point.