Add EntityType and EntitySubType definitions
Add the EntityType and EntitySubType definitions for Military Fixed
wing. These fields are important to OverlordBot so defining them first
for testing purposes.

Because we cannot share Enums or have Enum Interfaces or have any sort
of Enum inheritance we are doing the only real option which is to define
them within their own classes and we will have to do shennanigans when
retrieving the values to cast them to their appropriate enum.