Periodically write airbases to the Database
Periodically write anything that counts as an Airbase, according to DCS,
to the database.

Since this happens only once per minute and the number of units should
never be very high we will just truncate and write everything again. The
performance impact of this will be minimal given the number of units and
it reduces the amount of complexity required in the code.