Make tasks configurable in the configuration
Make it possible to specify if tasks are enabled or not and their
settings using a configuration file to make it easier for users to
decide what features they want or do not want.

All time values in the configuration are in seconds. We may want
to switch the MarkPanel task to read from the event stream and
write to the DB when changes occur instead of periodically getting
all Markpoints.